fix: set fake address for venue in event
We use the same model for creating an event as we do for viewing it. To view, we retrieve the first image url for that event. However, we obviously don't need to set that when we create the event.

The request fails if the value is not a valid url. It's a bit hacky, but adding a fake default address fixes this.

The post request doesn't use this information at all - so no harm done.